modal tested and integrated on tripCard
