--- client/src/components/modal/Modal_ejemplo.css
@@ -1,20 +1,0 @@
-.modal {
-    position: fixed;
-    margin:0;
-    width:100%;
-    height:100%;
-    left:0;
-    top:0;
-    background-color: rgba(0,0,0,0.6);
-    display: flex;
-}
-.modal-body {
-    background-color: white;
-    width:50%;
-    margin: auto;
-    min-height:50%;
-    border-radius:15px;
-    padding:2rem;
-    position: relative;
-
-}
