+++ client/src/components/modal/modal.css
@@ -1,57 +1,20 @@
-.modal {   
-    background-color: rgba(0, 0, 0, 0.5);
+.modal {
     position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100vh;
+    margin:0;
+    width:100%;
+    height:100%;
+    left:0;
+    top:0;
+    background-color: rgba(0,0,0,0.6);
     display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .modal-container {
-    background-color: black;
-    width: 50%;
-    min-width: 400px;
-    border-radius: 1rem;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.5);
+}
+.modal-body {
+    background-color: white;
+    width:50%;
+    margin: auto;
+    min-height:50%;
+    border-radius:15px;
+    padding:2rem;
     position: relative;
-    padding: 1rem;
-  }
-  
-  .modal-close {
-    position: absolute;
-    top: 0;
-    right: 0;
-    margin: 0.5rem;
-    background: none;
-    border: none;
-    font-size: 1.5rem;
-    cursor: pointer;
-  }
-  
-  h1 {
-    text-align: center;
-    margin: 1rem 0;
-    font-weight: bold;
-    font-size: 1.5rem;
-    color: black;
-  }
-  
-  .modal-body {
-    display: flex;
-    align-items: center;
-  }
-  
-  .miImagenId {
-    width: 35%;
-    height: auto;
-    margin-right: 1rem;
-  }
-  
-  .modal-instructions {
-    flex: 1;
-    text-align: center;
-  }
-  
+
+}
